Build for AArch64 (Requires update to Eclipse 12-2021)
Once upgraded to Eclipse 12-2021 all the existing code just worked.
Tycho is not building the update site properly for me but all tests pass
and the generated launcher.macos and launcher.core jars load and work
fine if copied into an eclipse plugins folder.

Upgrade to build for Eclipse 12-2021 required:
* Updated to SWTBOT 3.1.0
* Use Java 11 build environment not 1.8
* MaxPermSize is no longer an option in Java 11
* Newer Tycho Plugin to support Java 11